refactor(server): extract /connections/* routes to connectorRoutes.ts
Mechanical lift of the 1090-line connector-routes block out of server.ts
(per the 2026-04-30 tech-debt assessment, slice 1 of the server.ts split).
server.ts shrinks 3579 → 2495 lines (-30%); connector dispatch becomes one
line at the call site:

    if (tryHandleConnectorRoute(req, res, parsedUrl)) return;

No behavior change. The handler bodies are byte-identical to the originals
save for wrapping ~10 routes (slack-auth, slack/notion/confluence/zendesk/
intercom-disconnect, hubspot-test/disconnect, datadog-disconnect, stripe-
test/disconnect) in `void (async()=>{...})()` so the new module can return
boolean synchronously rather than Promise<boolean>. The microtask delay
this introduces is invisible to clients — the parent request handler
returns immediately on a true result either way, and res.end() has always
been async.

Verified zero closure deps before extraction: the 1090-line block contains
zero `this.` references, so no dependency injection was needed — the new
module imports nothing from server.ts.

Out of scope (deferred to follow-up): the 7 connector OAuth-callback routes
that live up at server.ts lines 590-720 are NOT moved here. They run inside
the OAuth authentication flow with bearer-auth ordering constraints; safer
to consolidate them once this seam has soaked.

Verified:
- Full bridge test suite: 322 files / 4492 tests green (no regressions)
- TypeScript build clean
- Zero LSP diagnostics on both files
- Biome clean
